chore(BridgeMaker) : move BridgeMaker
