+++ src/main/java/bridge/BridgeMaker.java
@@ -1,4 +1,4 @@
-package bridge.Domain;
+package bridge;
 
 import bridge.BridgeNumberGenerator;
 import java.util.ArrayList;
