+++ src/test/java/bridge/DomainTest/BridgeMakerTest.java
@@ -2,7 +2,7 @@ package bridge.DomainTest;
 
 import bridge.BridgeNumberGenerator;
 import bridge.BridgeRandomNumberGenerator;
-import bridge.Domain.BridgeMaker;
+import bridge.BridgeMaker;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
